Friday 11/11 Update
Cleaned up some code. Changed the map to an object type 2d array. Populating the 2d array with Wall and Monster objects. Each has attributes that hold data for the room type/description, if contains a potion, or if it is a wall or exit. Unable to get it to hold a Monster Object. Experimenting with doing this backwards and instead of doing room types that contain a monster, placing the monsters at coordinates and giving them location descriptions.

Made methods to randomly spawn the map with a (reachable!) exit and randomly spawn the map with the Wizert.

Unsure if to make the Move method part of the Hero class or the Map class.