Question/Stuck
Major apologies as I have a ton of commented out code that is a product of me working on 3 different sets of ideas that are all in varying stages.

What I am trying to accomplish is having my 2d array populated with Room objects that are either walls or non walls. The non walls I am trying to have contain relevant data to trigger when the Wizert moves into it, such as room description and to begin a battle. In my Map class you can see me trying to create a room that holds a Monster object within it but I can not get that to work. That is what I am most stuck on and have exhausted myself trying to find a solution. Any direction or resources regarding this would be incredibly appreciated.